clang build

+++ C/7zTypes.h
@@ -224,8 +224,8 @@ typedef __int64 Int64;
 typedef unsigned __int64 UInt64;
 #define UINT64_CONST(n) n
 #else
-typedef long long int Int64;
-typedef unsigned long long int UInt64;
+typedef long int Int64;
+typedef unsigned long int UInt64;
 #define UINT64_CONST(n) n ## ULL
 #endif
